packages/linkml_runtime/requirements.txt
################################################################################
# This requirements file has been automatically generated from `Pipfile` with
# `pipenv-to-requirements`
#
#
# This has been done to maintain backward compatibility with tools and services
# that do not support `Pipfile` yet.
#
# Do NOT edit it directly, use `pipenv install [-d]` to modify `Pipfile` and
# `Pipfile.lock` and then regenerate `requirements*.txt`.
################################################################################

click
hbreader
jsonasobj
linkml-model
prefixcommons
pyld
pyyaml>=5.1
rdflib
rdflib-jsonld
rdflib-pyld-compat
shexjsg
